Reduce duplication in bugreport ps output.
PROCESSES AND THREADS and PROCESSES (SELINUX LABELS) can be combined.
I suspect we can just remove PROCESSES too, since it's a strict
subset of the other information, but one step at a time. (The worry
is that someone's actually parsing this output, but it would be good
to find out who's doing that as soon as possible so we can work with
them as we transition to a more complete ps implementation.)

Change-Id: I5f1085f8853f7a988f7edcaeefdd7b63175c1473